Infer version number from git tags thanks to setuptools_scm
To be consistent, we define decomon.__version__ by looking at metadata
of installed package. If not installed, __version__ remains undefined.
We need importlib-metadata for that which is by default on python 3.8
but must be installed for previous versions.

cf https://github.com/pypa/setuptools_scm#retrieving-package-version-at-runtime

The version will be inferred from distance to tags of the form
"v{x}.{y}.{z}" or "{x}.{y}.{z}" where {x}, {y}, and {z} are numbers.