Use helpers functions in tests without using tests/__init__.py
With the pytest option "--import-mode=importlib", tests is not anymore
in the python path. So we need to avoid considering as a package.

We implement a solution to fix this by using conftest.py file used by
pytest to share common fixtures. Actually, we follow the suggestion made
here: https://stackoverflow.com/a/42156088