Remove argument slope from wrapper
We undo adding this argument.
The idea is that if we let the user decide the slope there, we should
also let it specify finetune, method, mode, and so on.
Instead we use default parameters for clone/convert. If the use wants to
specify the way to convert the model, he can still do the conversion of
the keras model himself before passing to the wrapper functions, as it
is already explained in tutorials.