Make to_decomon returns a single layer
We can now assume that activations are put in separate layers as it is
done at the beginning of convert().

So there is no need anymore to return an additional DecomonActivation
when converting Dense or Conv2D layers.

But we still need convert_forward_functional_model() to take a layer_fn
returning a list to split layers with non-linear activation.

Tests have also to be modified accordingly. We add tests for failure
cases of to_decomon.