Remove duplicate code initializing input_tensors for convert
clone() function already initialize input_tensors so we can assume that
input_tensors is not None in convert_forward and convert_backward.

We replace the implementation of get_input_tensors() by the one used in
clone() and call it within it.

We remove not anymore used check_input_tensors_sequential()