Fix way to get layer input shape in to_decomon
In order to init input_tensors to have a first a call on a new decomon layer,
we used layer.input_shape.

However this is not working for submodels after the preprocess pass.
Because the input layer of the submodel has now 2 inbound nodes, the
original one and the one constructed during preprocess, and the new inbound
node express its input shapes as a list (of one input shape) instead of
a single input shape and keras Layer original function thinks that the 2
input shapes differ.

We define a specific function to extract the input shape which first
transform a single input shape into a list with one element (the input
shape itself).

This fixes the issue.