Replace layer.input_shape by layer.input.shape
layer.input_shape does not exist anymore.
We can look for layer.input instead.
And we do not need to raise an error if not existing as there is already
an error raised by keras with a pertinent message when the layer has not
been called before and thus has nothing to return for input property.
Howver error is now of type ValueError => change in tests