Force compute_output_spec() to use Layer.compute_output_spec()
The idea is to avoid overriden versions like in Flatten which make the
assumption that a single output shape is returned. The version in Layer
allows a list of shapes as we do in compute_output_shape()

We also update compute_output_shape to return a list of shapes, the same
as the inputs by default.