Fix new keras operations to work on symbolic backend tensors
When a layer has no compute_output_shape() defined (for now
backward layers), the output shape is inferred by calling the layers
with symbolic tensors as inputs. But they are not KerasTensor anymore
but rather backend tensors (ie f.Tensor) with batchsize equal to None.

In this case, keras Operation by default use the call() rather than
compute_output_spec() so we modify it to use rather compute_output_spec.