Update and simplify tox config file to ensure test coverage
Surprisingly, having the section [testenv:py39-linux-nodeellip] after
the more generic section
[testenv:py{39,311}-{linux,macos,win}-nodeellip] skip the more specific
section definition.

So the coverage was missed. This can be corrected by putting
[testenv:py39-linux-nodeellip] before [testenv:py{39,311}-{linux,macos,win}-nodeellip]
but this not intuitive.

We choose here to do everything in the main section [testenv] by making
full use of generative names and factor conditions.

In particular, to have a command be used if a given factor is absent, by
using an exclamation mark:

	!nodeellip: ...