Add a get_config() method for DecomonModel
Even though this is actually a Functional model, it does not use its
get_config() method because it is checked inside it that the __init__
method has the same signature as a functional model.

Thus we enforce the get_config() from Functional by converting the model
to a mere keras functional model and then add the properties specific to
decomon.

We add a test for it.