Robustify isort linter
Sometimes isort can set 3rd-party libaries as 1st-party ones if a random
directory with the library name in the project arborescence
(on the repository or even locally). And thus make `pre-commit`
change wrongly a file.

To avoid this we specify to isort where the source directory for our
library is.