Share weights when splitting activation
We use the new utilty function to share weights between layer and layer
without activation so that further update of the weights in the original
model impact the split model.

NB: we make the assumption in the utility function that the name of the
weights are also the names of the corresponding layer attributes. This
seems not to be the case for deel-lip (e.g. SpectralDense.u is named
"sn"). For now the library is not keras 3 ready, but when we support
again deel-lip, some modifications will be necessary to take care of it.
Maybe a feature request to name them according to the attribute name, on
deel-lip side.