Fix interactive plots on colab
We need to add `plt.show()` at the end of the plotting functions so that
the new plots replace the previous ones when moving a slider. Else they
are only displayed below the others.

The issue was not seen on local jupyter or binder but this fix is ok for
these platforms.