Fix weights sharing for Dense with kernel being an unsettable property
Since keras 3.0.3, kernel attribute is a *non-settable*  property,
in order to manage LoRA implementation (see
https://github.com/keras-team/keras/pull/18942)

So sharing the kernel requires to set layer._kernel instead of kernel.
This is a bit hacky, but no easy turn-around with the way decomon is
designed for now.

A full refactoring is coming that may simplify the weights sharing
(replaced then by a layer sharing).

For now we try setting the attribute with the same name prefixed by "_"
when it is impossible to set the attribute with weight name itself.
Namely, we set layer._kernel instead of layer.kernel.