Add batch_multid_dot() operation on tensors
This is a multid equivalent of batch_dot() on tensors.
We perform a dot product on tensors by summing over a range of axes
instead of a single axis.

In the first tensor, we perform it on the last axes, starting from a
given one.
In the second tensor, we perform it on the first axes (after batch
axis), ending to a given one.

The option `missing_batchsize` is used to apply batch_multi_dot even
when the batch axis is missing for one of the entry. (In that case, we
use keras.ops.tensordot under the hood.)
This can be useful when combining an affine bounds having a batch size
with a layer affine representation without the batch dimension.

By default, the number of axes to merge is the number of non-batch axes
in the first arg, so that a linear operator represented by tensor `w`
operates on an input tensor `x` as `batch_multi_dot(x, w,
missing_batchsize=(False, True)`