Update perturbation domain to manage empty/diagonal/no batch affine bounds
For affine bounds empty (ie identity), we add a get_lower_x and
get_upper_x method that return lower and upper bounds on the model input
x.

We implement it in box case.
We also take car of diagonal/ w/o batchsize inputs in box case.

Ball perturbation domains are to be completed later.