Fix convert_backward with embedded submodels + full crown
In full crown mode, to get oracle bounds inside an embedded submodel,
we need to reconstruct the proper perturbation domain inputs for this
submodel.

We do it dynamically when necessary (to avoid it if the submodel has
only linear layers for instance).

We deduce it from constant bounds on submodel input, i.e. oracle bounds
for the node corresponding to the submodel.

To reconstruct from them the perturbation domain input, we need a new
method PerturbationDomain.get_input_from_constant_bounds() which
generates x from lower and upper.

We reuse the formula used in previous version of the code which was
valid only for BoxDomain.

Not yet implemented for BallDomain.