Fix PerturbationDomain for affine bounds w/o batchsize
The way  we checked for affine bounds w/o batchsize was wrong (we
compared b and x shape although b is of the output shape and x of the
input shape).
We need actually a specific arg to specify whether w and b are missing
batchsize or not.

We also add doctrings to methods for perturbation domains.