Fix renaming of reshaped affine inputs
The affine bounds are reshaped to be passed through self.layer
but also renamed and sometimes the new name is "forgotten" afterwards
(in particular in DecomonMerge).
So we averwrite the original b_l_in, w_l_in, ... instead to avoid this.

This fixes some issues in test_clone for model add_linear.