Fix get_linear_hull_s_shape
When using K.sign(lower_flat), one must take into account the fact that
it returns 0 when lower_flat is null.
This fact leads to alpha_u_1 being sometimes = 0.5 which could be
harmful.
We replace it by K.sign(K.sign(lower_flat) + 1.) which is 0 for negative
values and 1 for non-negative values.